requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioredis==1.3.1
aiosignal==1.3.2
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
asgiref==3.8.1
astroid==2.15.8
asttokens==3.0.0
async-timeout==4.0.3 ## async-timeout==5.0.1
atomicwrites==1.4.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
backcall==0.2.0
bcrypt==4.2.1
billiard==4.2.1
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.5.1
captcha==0.6.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.3.1
crashtest==0.4.1
cryptography==44.0.0
cycler==0.12.1
daphne==3.0.2
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
deepl==1.20.0
dill==0.3.9
distlib==0.3.9
distro==1.9.0
Django==5.1.4
django-extensions==3.2.3
dulwich==0.21.7
executing==2.1.0
faiss-cpu==1.9.0.post1
fastjsonschema==2.21.1
feedparser==6.0.11
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
hiredis==3.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
incremental==24.7.2
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.29.5
ipython==7.34.0
isort==5.13.2
jaraco.classes==3.4.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==24.3.1
kiwipiepy==0.20.3
kiwipiepy_model==0.20.0
kiwisolver==1.4.8
kombu==5.4.2
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-teddynote==0.3.41
langchain-text-splitters==0.3.4
langgraph==0.2.60
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.48
langsmith==0.2.7
lazy-object-proxy==1.10.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mmh3==4.1.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.25.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.58.1
orjson==3.10.13
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathtools==0.1.2
pdf2image==1.17.0
pdfminer.six==20240706
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.9.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==4.25.5
protoc-gen-openapiv2==0.0.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
pylint==2.17.7
PyMuPDF==1.25.2
pyOpenSSL==24.3.0
pyparsing==3.2.1
pypdf==5.1.0
PyPDF2==3.0.1
pyproject_hooks==1.2.0
pytest==6.2.5
pytest-django==4.5.2
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
rag==0.0.20
ragclip==0.0.4
rank-bm25==0.2.2
RapidFuzz==3.11.0
redis==3.5.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
sentence-transformers==3.3.1
service-identity==24.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
tavily-python==0.5.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.0
trove-classifiers==2024.10.21.16
Twisted==24.11.0
txaio==23.1.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
virtualenv==20.28.1
# watchdog==0.10.3
wcwidth==0.2.13
wget==3.2
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
zope.interface==7.2
